-Blaster remove term 'fork()' error in error message -Leave blast log when parameter CleanTmpFile is at false

+++ src/blaster/BLRNCBIBlast.cpp
@@ -48,7 +48,7 @@ void BLRNCBIBlast::blast( int verbose )
     if (sys_return == -1) {
         std::cout << std::ifstream(query_filename + "_blast.log").rdbuf();
         std::ostringstream ostr;
-        ostr << " call to 'system()' function return 'fork()' error -1 !";
+        ostr << " call to 'system()' function return error -1 !";
         throw SDGException(NULL, ostr.str(), -1);
     }
 
@@ -63,8 +63,10 @@ void BLRNCBIBlast::blast( int verbose )
 
     SDGString rm_command="rm -f "+query_filename;
     system( rm_command );
-    rm_command="rm -f "+query_filename+ "_blast.log";
-    system( rm_command );
+    if(para.getCleanTmpFiles()){
+        rm_command="rm -f "+query_filename+ "_blast.log";
+        system( rm_command );
+    }
 }
 
 void BLRNCBIBlast::pressdb( int verbose )
